organization for external parallelism
