+++ perses/dispersed/smc.py
@@ -27,6 +27,7 @@ import time
 from collections import namedtuple
 from perses.annihilation.lambda_protocol import LambdaProtocol
 from perses.annihilation.lambda_protocol import RelativeAlchemicalState, LambdaProtocol
+from perses.dispersed import *
 import random
 import pymbar
 import dask.distributed as distributed
@@ -40,167 +41,6 @@ _logger.setLevel(logging.DEBUG)
 
 #cache.global_context_cache.platform = openmm.Platform.getPlatformByName('Reference') #this is just a local version
 EquilibriumFEPTask = namedtuple('EquilibriumInput', ['sampler_state', 'inputs', 'outputs'])
-
-class DaskClient(object):
-    """
-    This class manages the dask scheduler.
-    Parameters
-    ----------
-    LSF: bool, default False
-        whether we are using the LSF dask Client
-    num_processes: int, default 2
-        number of processes to run.  If not LSF, this argument does nothing
-    adapt: bool, default False
-        whether to use an adaptive scheduler.  If not LSF, this argument does nothing
-    """
-    def __init__(self):
-        _logger.info(f"Initializing DaskClient")
-
-    def activate_client(self,
-                        LSF = True,
-                        num_processes = 2,
-                        adapt = False,
-                        timeout = 1800):
-
-        if LSF:
-            from dask_jobqueue import LSFCluster
-            cluster = LSFCluster()
-            self._adapt = adapt
-            self.num_processes = num_processes
-
-            if self._adapt:
-                _logger.debug(f"adapting cluster from 1 to {self.num_processes} processes")
-                cluster.adapt(minimum = 2, maximum = self.num_processes, interval = "1s")
-            else:
-                _logger.debug(f"scaling cluster to {self.num_processes} processes")
-                cluster.scale(self.num_processes)
-
-            _logger.debug(f"scheduling cluster with client")
-            self.client = distributed.Client(cluster, timeout = timeout)
-            while len(self.client.nthreads()) != self.num_processes:
-                _logger.debug(f"workers: {self.client.nthreads()}.  waiting for {self.num_processes} workers...")
-                time.sleep(3)
-            worker_threads = self.client.nthreads()
-            self.workers = {i: _worker for i, _worker in zip(range(len(worker_threads)), worker_threads.keys())}
-            self.worker_counter = 0
-
-            #now we wait for all of our workers.
-        else:
-            self.client = None
-            self._adapt = False
-            self.num_processes = 0
-
-    def deactivate_client(self):
-        """
-        NonequilibriumSwitchingFEP is not pickleable with the self.client or self.cluster activated.
-        This must be called before pickling
-        """
-        if self.client is not None:
-            self.client.close()
-            self.client = None
-            self.workers = None
-            self.worker_counter = 0
-
-    def scatter(self, df):
-        """
-        wrapper to scatter the local data df
-        """
-        if self.client is None:
-            #don't actually scatter
-            return df
-        else:
-
-            return self.client.scatter(df)
-
-    def deploy(self, func, arguments):
-        """
-        wrapper to map a function and its arguments to the client for scheduling
-        Arguments
-        ---------
-        func : function to map
-            arguments: tuple of the arguments that the function will take
-        argument : tuple of argument lists
-        Returns
-        ---------
-        futures
-        """
-        if self.client is None:
-            if len(arguments) == 1:
-                futures = [func(plug) for plug in arguments[0]]
-            else:
-                futures = [func(*plug) for plug in zip(*arguments)]
-        else:
-            futures = self.client.map(func, *arguments)
-        return futures
-
-    def gather_results(self, futures):
-        """
-        wrapper to gather a function given its arguments
-        Arguments
-        ---------
-        futures : future pointers
-
-        Returns
-        ---------
-        results
-        """
-        if self.client is None:
-            return futures
-        else:
-            results = self.client.gather(futures)
-            return results
-
-    def gather_actor_result(self, future):
-        """
-        wrapper to pull the .result() of a method called to an actor
-        """
-        if self.client is None:
-            return future
-        else:
-            distributed.progress(future)
-            result = future.result()
-            return result
-
-    def launch_actor(self, _class):
-        """
-        wrapper to launch an actor
-
-        Arguments
-        ---------
-        _class : class object
-            class to put on a worker
-
-        Returns
-        ---------
-        actor : dask.distributed.Actor pointer (future)
-        """
-        if self.client is not None:
-            future = self.client.submit(_class, workers = [self.workers[self.worker_counter]], actor=True)  # Create a _class on a worker
-            self.worker_counter += 1
-            distributed.progress(future)
-            actor = future.result()                    # Get back a pointer to that object
-            return actor
-        else:
-            actor = _class()
-            return actor
-
-    def progress(self, futures):
-        """
-        wrapper to log the progress of futures
-        """
-        if self.client is None:
-            pass
-        else:
-            distributed.progress(futures)
-
-    def wait(self, futures):
-        """
-        wrapper to wait until futures are complete.
-        """
-        if self.client is None:
-            pass
-        else:
-            distributed.wait(futures)
 
 class SequentialMonteCarlo():
     """
@@ -227,9 +67,10 @@ class SequentialMonteCarlo():
                  ncmc_save_interval = None,
                  measure_shadow_work = False,
                  neq_integrator = 'langevin',
-                 LSF = False,
-                 num_processes = 2,
-                 adapt = False):
+                 external_parallelism = None,
+                 internal_parallelism = {'library': ('dask', 'LSF'),
+                                         'num_processes': 2}
+                                         ):
         """
         Parameters
         ----------
@@ -261,12 +102,11 @@ class SequentialMonteCarlo():
             WARNING : this is not currently supported
         neq_integrator : str, default 'langevin'
             which integrator to use
-        LSF: bool, default False
-            whether we are using the LSF dask Client
-        num_processes: int, default 2
-            number of processes to run.  If not LSF, this argument does nothing
-        adapt: bool, default False
-            whether to use an adaptive scheduler.  If not LSF, this argument does nothing
+        external_parallelism : dict('parallelism': perses.dispersed.parallel.Parallelism, 'available_workers': list(str)), default None
+            an external parallelism dictionary
+        internal_parallelism : dict, default {'library': ('dask', 'LSF'), 'num_processes': 2}
+            dictionary of parameters to instantiate a client and run parallel computation internally.  internal parallelization is handled by default
+            if None, external worker arguments have to be specified, otherwise, no parallel computation will be conducted, and annealing will be conducted locally.
         """
         _logger.info(f"Initializing SequentialMonteCarlo")
 
@@ -358,9 +198,11 @@ class SequentialMonteCarlo():
         self.sampler_states = {0: copy.deepcopy(sampler_state), 1: copy.deepcopy(sampler_state)}
 
         #Dask implementables
-        self.LSF = LSF
-        self.num_processes = num_processes
-        self.adapt = adapt
+        self.external_parallelism = external_parallelism
+        self.internal_parallelism = internal_parallelism
+        if external_parallelism is not None and internal_parallelism is not None:
+            raise Exception(f"external parallelism were given, but an internal parallelization scheme was also specified.  Aborting!")
+
 
     def launch_LocallyOptimalAnnealing(self):
         """
@@ -455,100 +297,100 @@ class SequentialMonteCarlo():
 
         return num_actors, num_actors_per_direction, num_particles_per_actor, sMC_actors
 
-    def AIS(self,
-            num_particles,
-            protocol_length,
-            directions = ['forward','reverse'],
-            num_integration_steps = 1,
-            return_timer = False,
-            rethermalize = False):
-        """
-        Conduct vanilla AIS. with a linearly interpolated lambda protocol
-
-        Arguments
-        ----------
-        num_particles : int
-            number of particles to run in each direction
-        protocol_length : int
-            number of lambdas
-        directions : list of str, default ['forward', 'reverse']
-            the directions to run.
-        num_integration_steps : int
-            number of integration steps per proposal
-        return_timer : bool, default False
-            whether to time the annealing protocol
-        rethermalize : bool, default False
-            whether to rethermalize velocities after proposal
-        """
-        self.activate_client(LSF = self.LSF,
-                            num_processes = self.num_processes,
-                            adapt = self.adapt)
-
-        for _direction in directions:
-            assert _direction in ['forward', 'reverse'], f"direction {_direction} is not an appropriate direction"
-        protocols = {}
-        for _direction in directions:
-            if _direction == 'forward':
-                protocol = np.linspace(0, 1, protocol_length)
-            elif _direction == 'reverse':
-                protocol = np.linspace(1, 0, protocol_length)
-            protocols.update({_direction: protocol})
-
-        num_actors, num_actors_per_direction, num_particles_per_actor, AIS_actors = self._actor_distribution(directions = directions,
-                                                                                                             num_particles = num_particles)
-        AIS_actor_futures = {_direction: [[[None] * num_particles_per_actor] for _ in range(num_actors_per_direction)]}
-        AIS_actor_results = copy.deepcopy(AIS_actor_futures)
-
-        for job in tqdm.trange(num_particles_per_actor):
-            start_timer = time.time()
-            for _direction in directions:
-                _logger.info(f"entering {_direction} direction to launch annealing jobs.")
-                for _actor in range(num_actors_per_direction):
-                    _logger.info(f"\tretrieving actor {_actor}.")
-                    sampler_state = self.pull_trajectory_snapshot(0) if _direction == 'forward' else self.pull_trajectory_snapshot(1)
-                    if self.ncmc_save_interval is not None: #check if we should make 'trajectory_filename' not None
-                        noneq_trajectory_filename = self.neq_traj_filename[_direction] + f".iteration_{self.total_jobs:04}.h5"
-                        self.total_jobs += 1
-                    else:
-                        noneq_trajectory_filename = None
-
-                    actor_future = AIS_actors[_direction][_actor].anneal(sampler_state = sampler_state,
-                                                                         lambdas = protocols[_direction],
-                                                                         label = self.total_jobs,
-                                                                         noneq_trajectory_filename = noneq_trajectory_filename,
-                                                                         num_integration_steps = num_integration_steps,
-                                                                         return_timer = return_timer,
-                                                                         return_sampler_state = False,
-                                                                         rethermalize = rethermalize,
-                                                                         )
-                    AIS_actor_futures[_direction][_actor][job] = actor_future
-
-
-            #now we collect the jobs
-            for _direction in directions:
-                for _actor in range(num_actors_per_direction):
-                    _future = AIS_actors_futures[_direction][_actor][job]
-                    result = self.gather_actor_result(_future)
-                    AIS_actor_results[_direction][_actor][job] = result
-
-            end_timer = time.time() - start_timer
-            _logger.info(f"iteration took {end_timer} seconds.")
-
-
-        #now that the actors are gathered, we can collect the results and put them into class attributes
-        _logger.info(f"organizing all results...")
-        for _direction in AIS_actor_results.keys():
-            _result_lst = AIS_actor_results[_direction]
-            _logger.info(f"collecting {_direction} actor results...")
-            flattened_result_lst = [item for sublist in _result_lst for item in sublist]
-            [self.incremental_work[_direction].append(item[0]) for item in flattened_result_lst]
-            [self.nonequilibrium_timers[_direction].append(item[2]) for item in flattened_result_lst]
-
-        #compute the free energy
-        self.compute_free_energy()
-
-        #deactivate_client
-        self.deactivate_client()
+    # def AIS(self,
+    #         num_particles,
+    #         protocol_length,
+    #         directions = ['forward','reverse'],
+    #         num_integration_steps = 1,
+    #         return_timer = False,
+    #         rethermalize = False):
+    #     """
+    #     Conduct vanilla AIS. with a linearly interpolated lambda protocol
+    #
+    #     Arguments
+    #     ----------
+    #     num_particles : int
+    #         number of particles to run in each direction
+    #     protocol_length : int
+    #         number of lambdas
+    #     directions : list of str, default ['forward', 'reverse']
+    #         the directions to run.
+    #     num_integration_steps : int
+    #         number of integration steps per proposal
+    #     return_timer : bool, default False
+    #         whether to time the annealing protocol
+    #     rethermalize : bool, default False
+    #         whether to rethermalize velocities after proposal
+    #     """
+    #     self.activate_client(LSF = self.LSF,
+    #                         num_processes = self.num_processes,
+    #                         adapt = self.adapt)
+    #
+    #     for _direction in directions:
+    #         assert _direction in ['forward', 'reverse'], f"direction {_direction} is not an appropriate direction"
+    #     protocols = {}
+    #     for _direction in directions:
+    #         if _direction == 'forward':
+    #             protocol = np.linspace(0, 1, protocol_length)
+    #         elif _direction == 'reverse':
+    #             protocol = np.linspace(1, 0, protocol_length)
+    #         protocols.update({_direction: protocol})
+    #
+    #     num_actors, num_actors_per_direction, num_particles_per_actor, AIS_actors = self._actor_distribution(directions = directions,
+    #                                                                                                          num_particles = num_particles)
+    #     AIS_actor_futures = {_direction: [[[None] * num_particles_per_actor] for _ in range(num_actors_per_direction)]}
+    #     AIS_actor_results = copy.deepcopy(AIS_actor_futures)
+    #
+    #     for job in tqdm.trange(num_particles_per_actor):
+    #         start_timer = time.time()
+    #         for _direction in directions:
+    #             _logger.info(f"entering {_direction} direction to launch annealing jobs.")
+    #             for _actor in range(num_actors_per_direction):
+    #                 _logger.info(f"\tretrieving actor {_actor}.")
+    #                 sampler_state = self.pull_trajectory_snapshot(0) if _direction == 'forward' else self.pull_trajectory_snapshot(1)
+    #                 if self.ncmc_save_interval is not None: #check if we should make 'trajectory_filename' not None
+    #                     noneq_trajectory_filename = self.neq_traj_filename[_direction] + f".iteration_{self.total_jobs:04}.h5"
+    #                     self.total_jobs += 1
+    #                 else:
+    #                     noneq_trajectory_filename = None
+    #
+    #                 actor_future = AIS_actors[_direction][_actor].anneal(sampler_state = sampler_state,
+    #                                                                      lambdas = protocols[_direction],
+    #                                                                      label = self.total_jobs,
+    #                                                                      noneq_trajectory_filename = noneq_trajectory_filename,
+    #                                                                      num_integration_steps = num_integration_steps,
+    #                                                                      return_timer = return_timer,
+    #                                                                      return_sampler_state = False,
+    #                                                                      rethermalize = rethermalize,
+    #                                                                      )
+    #                 AIS_actor_futures[_direction][_actor][job] = actor_future
+    #
+    #
+    #         #now we collect the jobs
+    #         for _direction in directions:
+    #             for _actor in range(num_actors_per_direction):
+    #                 _future = AIS_actors_futures[_direction][_actor][job]
+    #                 result = self.gather_actor_result(_future)
+    #                 AIS_actor_results[_direction][_actor][job] = result
+    #
+    #         end_timer = time.time() - start_timer
+    #         _logger.info(f"iteration took {end_timer} seconds.")
+    #
+    #
+    #     #now that the actors are gathered, we can collect the results and put them into class attributes
+    #     _logger.info(f"organizing all results...")
+    #     for _direction in AIS_actor_results.keys():
+    #         _result_lst = AIS_actor_results[_direction]
+    #         _logger.info(f"collecting {_direction} actor results...")
+    #         flattened_result_lst = [item for sublist in _result_lst for item in sublist]
+    #         [self.incremental_work[_direction].append(item[0]) for item in flattened_result_lst]
+    #         [self.nonequilibrium_timers[_direction].append(item[2]) for item in flattened_result_lst]
+    #
+    #     #compute the free energy
+    #     self.compute_free_energy()
+    #
+    #     #deactivate_client
+    #     self.deactivate_client()
 
     def sMC_anneal(self,
                    num_particles,
